requirements.txt
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.2
drf-yasg2==1.19.1
gunicorn==20.0.4
psycopg2==2.8.6
python-dotenv==0.15.0 
requests==2.25.0
requests-html==0.10.0
rq==1.7.0
six==1.15.0
whitenoise==5.2.0
